docs/hug25/pyproject.toml
[tool.poetry]
name = "pyconau23"
version = "0.1.0"
description = ""
authors = ["Clinton Roy <clintonr@ansto.gov.au>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
h5py = "^3.9.0"
h5pydantic = {git = "https://gitlab.com/clinton.roy/h5pydantic.git"}
pint = "^0.22"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
